Add free floating stations management
