+++ examples/vehicle_sharing/mlgraph.json
@@ -492,54 +492,54 @@
       "RES": {
         "id": "RES",
         "sections": [
-          "WEST_0_0",
-          "0_WEST_0",
-          "WEST_2_2",
-          "WEST_1_1",
-          "4_5",
-          "6_3",
-          "3_SOUTH_1",
-          "1_0",
-          "7_4",
-          "NORTH_2_8",
+          "6_EAST_0",
           "1_WEST_1",
-          "8_5",
-          "EAST_1_7",
-          "4_3",
+          "EAST_0_6",
           "1_4",
-          "0_SOUTH_0",
-          "3_0",
-          "5_NORTH_1",
-          "2_WEST_2",
           "SOUTH_1_3",
-          "7_EAST_1",
-          "NORTH_1_5",
-          "7_6",
-          "8_7",
+          "4_5",
+          "WEST_2_2",
           "2_1",
+          "2_WEST_2",
+          "3_SOUTH_1",
+          "0_3",
+          "1_2",
+          "1_0",
+          "NORTH_0_2",
+          "8_5",
+          "7_8",
+          "SOUTH_0_0",
+          "3_0",
+          "8_EAST_2",
+          "5_NORTH_1",
+          "SOUTH_2_6",
           "5_4",
+          "8_7",
+          "3_6",
+          "5_2",
+          "4_3",
+          "2_5",
+          "4_7",
+          "5_8",
+          "8_NORTH_2",
+          "NORTH_2_8",
+          "0_SOUTH_0",
+          "7_EAST_1",
+          "EAST_2_8",
+          "NORTH_1_5",
+          "6_3",
+          "6_SOUTH_2",
+          "6_7",
+          "0_WEST_0",
+          "7_4",
+          "EAST_1_7",
+          "4_1",
+          "2_NORTH_0",
           "0_1",
           "3_4",
-          "6_7",
-          "SOUTH_0_0",
-          "2_5",
-          "EAST_0_6",
-          "0_3",
-          "3_6",
-          "4_1",
-          "8_EAST_2",
-          "NORTH_0_2",
-          "5_8",
-          "4_7",
-          "EAST_2_8",
-          "2_NORTH_0",
-          "8_NORTH_2",
-          "1_2",
-          "6_SOUTH_2",
-          "SOUTH_2_6",
-          "5_2",
-          "7_8",
-          "6_EAST_0"
+          "WEST_1_1",
+          "7_6",
+          "WEST_0_0"
         ],
         "contour": [
           [
